Update the implementation to be user-specific login system

+++ backend/src/models/tripModel.js
@@ -1,6 +1,6 @@
 const pool = require("../config/db");
 
-const getAllTrips = async () => {
+const getAllTrips = async (user_id) => {
   const result = await pool.query(
     `SELECT
      id,
@@ -12,13 +12,16 @@ const getAllTrips = async () => {
      currency,
      num_of_people,
      created_at
-     FROM trips ORDER BY created_at DESC`
+     FROM trips 
+     WHERE user_id = $1
+     ORDER BY created_at DESC`,
+     [user_id]
   );
 
   return result.rows;
 };
 
-const getTripById = async (id) => {
+const getTripById = async (id, user_id) => {
   const result = await pool.query(
     `SELECT id, 
     title, 
@@ -30,8 +33,9 @@ const getTripById = async (id) => {
     num_of_people, 
     created_at 
     FROM trips 
-    WHERE id = $1`
-    , [id]);
+    WHERE id = $1
+      AND user_id = $2`
+    , [id, user_id]);
   return result.rows[0];
 };
 
@@ -61,7 +65,7 @@ const createTrip = async(tripData) => {
   return result.rows[0];
 };
 
-const updateTrip = async(id, tripData) => {
+const updateTrip = async(id, user_id, tripData) => {
   const {
     title, destination, start_date, end_date, total_budget, currency, num_of_people,
   } = tripData;
@@ -75,24 +79,24 @@ const updateTrip = async(id, tripData) => {
       total_budget = $5, 
       currency = $6,
       num_of_people = $7
-     WHERE id = $8
+     WHERE id = $8 AND user_id = $9
      RETURNING *`,
-    [ title, destination, start_date, end_date, total_budget, currency, num_of_people, id,]
+    [ title, destination, start_date, end_date, total_budget, currency, num_of_people, id, user_id,]
   );
 
   return result.rows[0];
 };
 
-const deleteTrip = async (id) => {
+const deleteTrip = async (id, user_id) => {
   const result = await pool.query(
-    `DELETE FROM trips WHERE id = $1 RETURNING *`,
-    [id]
+    `DELETE FROM trips WHERE id = $1 AND user_id = $2 RETURNING *`,
+    [id, user_id]
   );
 
   return result.rows[0];
 };
 
-const updateTripPeopleCount = async (tripId) => {
+const updateTripPeopleCount = async (tripId, user_id) => {
   const result = await pool.query(
     `UPDATE trips
      SET num_of_people = (
@@ -101,8 +105,9 @@ const updateTripPeopleCount = async (tripId) => {
       WHERE trip_id = $1
      )
      WHERE id = $1
+      AND user_id = $2
      RETURNING *`,
-    [tripId]
+    [tripId, user_id]
   );
   return result.rows[0];
 };
